feat: split createCompareFn function from is same type
