+++ src/create-compare-fn.ts
@@ -1,0 +1,45 @@
+import type { ZodType } from "zod";
+import { type CompareContext, type CompareRule } from "./types.ts";
+
+export const createCompareFn = (rules: CompareRule[]) => {
+  const isSameTypeFn = (
+    a: ZodType,
+    b: ZodType,
+    context: CompareContext = {},
+  ): boolean => {
+    let prevIndex = -1;
+    const runner = (index: number): boolean => {
+      if (index === rules.length) {
+        throw new Error("Failed to compare type! " + a + " " + b);
+      }
+      if (index === prevIndex) {
+        throw new Error("next() called multiple times");
+      }
+      prevIndex = index;
+      const rule = rules[index];
+
+      if ("stacks" in context && Array.isArray(context.stacks)) {
+        context.stacks.push({
+          name: rule.name,
+          target: [a, b],
+        });
+      }
+
+      return rule.compare(
+        a,
+        b,
+        () => runner(index + 1),
+        (a, b) => isSameTypeFn(a, b, context),
+        context,
+      );
+    };
+
+    return runner(0);
+  };
+  return isSameTypeFn;
+};
+
+/**
+ * @deprecated Use {@link createCompareFn} instead.
+ */
+export const createIsSameTypeFn = createCompareFn;
